Refactor CLI to use Cobra framework
Replace manual os.Args switching and flag.FlagSet with Cobra commands.
Moves each command into cmd/ package (root.go, serve.go, run.go);
main.go becomes a single cmd.Execute() call.
